feat: mixed-precision PixelDiT (fp32 master weights, bf16 compute)
PixelDiTParams gains a dtype=bf16 compute knob (param_dtype default
flips bf16->fp32); threaded through PatchTokenEmbedder, PixelTokenEmbedder,
CondTokenEmbedder, TimestepConditioner, SwiGLU, PixelMLP, MMDiTBlock and
PiTBlock: compute Linears/attention get dtype, norms stay fp32 islands,
FinalLayer stays an all-fp32 emit endpoint. Inputs (obs/cond/t) are cast
to fp32 at the model door.

Fixes a residual-stream leak: CondTokenEmbedder's RMSNorm (fp32 island)
returned its raw fp32 output directly into the cond-token stream instead
of downcasting first, re-promoting the whole y-stream back to fp32 via
JAX's bf16+fp32 promotion on every subsequent residual add -- the same
failure mode as the earlier flux1joint condition_embedding bug. Caught by
a new class-level monkeypatch regression test on MMDiTBlock.__call__.

Also pins dtype=jnp.float32 in two existing unit tests
(test_pixelmlp_uses_erf_gelu, test_pixel_token_embedder_pos_orientation_nonsquare)
that assert numeric equality against fp32 references tighter than bf16's
compute knob default now supports; that's a testing-methodology fix, not
a model behavior change.

+++ tests/experimental/models/pixeldit/test_embedders.py
@@ -225,13 +225,16 @@ def test_pixel_token_embedder_pos_orientation_nonsquare():
     from gensbi.experimental.models.pixeldit.modules import get_2d_sincos_pos_embed
 
     H, W, C, p, D_pix = 4, 6, 1, 2, 32
+    # dtype=jnp.float32 pins full-precision compute: this test verifies pos-
+    # embedding orientation math against an exact fp32 reference table, not
+    # mixed-precision behavior (bf16 compute would round entries visibly).
     m = PixelTokenEmbedder(
         C, D_pix, field_shape=(H, W), patch_size=p,
-        use_abs_pos=True, rngs=make_rngs(), param_dtype=jnp.float32,
+        use_abs_pos=True, rngs=make_rngs(), dtype=jnp.float32, param_dtype=jnp.float32,
     )
     m_no = PixelTokenEmbedder(
         C, D_pix, field_shape=(H, W), patch_size=p,
-        use_abs_pos=False, rngs=make_rngs(), param_dtype=jnp.float32,
+        use_abs_pos=False, rngs=make_rngs(), dtype=jnp.float32, param_dtype=jnp.float32,
     )
     # Share projection weights so we can isolate the pos contribution.
     m_no.proj = m.proj
